+++ guagua-core/src/main/java/ml/shifu/guagua/util/DiskList.java
@@ -190,8 +190,8 @@ public class DiskList<T extends Serializable> implements AppendList<T> {
     }
 
     @Override
-    public long size() {
-        return this.count;
+    public int size() {
+        return (int)this.count;
     }
 
     /*
